Refactor TransactionEndToEndTest to use RestAssured
The TransactionEndToEndTest was refactored to use the RestAssured library instead of the previously used MockMvc for performing HTTP requests in tests. This adjustment simplifies the tests by using the 'given' idiom consistently throughout the code and replacing multiple imports with the appropriate static imports. In addition, redundant and unused code was removed improving readability and maintainability of the test suite.